requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-socketio==5.3.6
mysql-connector-python==8.1.0
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
eventlet==0.33.3
pytest==7.4.0
python-dotenv==1.0.0
python-dateutil==2.8.2
